Implemented the exchange action and everything needed for it. It now works correctly. I fixed the SelectionMenu with allows for Selectables to be selected. I made an exchange menu which contains two selection menus for the entity inventories. It also contains a new class NumberBox which allows for numbers to be input, using buttons to change the number and a text line to show the display. I fixed a bug with TextBox which prevented the text from being changed. I implemented the code of the ExchangeAction, it opens and populates the ExchangeMenu if needed, then swapps the items in the inventories. I fixed loads of bugs with this action. I also discovered a worrying bug to do with the Menus not being displayed correctly after a certain amount of time, due to the code running out of memory to store the graphics. This is not yet fixed.

+++ src/Item.cpp
@@ -1,14 +1,14 @@
 #include "Item.h"
 #include "TextLine.h"
 
-Item::Item(int type, int amount, float pos_x, float pos_y)
+Item::Item(int type, float amount, float pos_x, float pos_y)
   : Selectable()
-{  
+{
   setPosX(pos_x);
   setPosY(pos_y);
 
   mSprite = new Sprite(pos_x, pos_y);
-  
+
   setType(type);
   setAmount(amount);
